requirements.txt
beautifulsoup4==4.12.2
Requests==2.31.0
customtkinter=5.2.0
Pillow=10.0.0